requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
httpx==0.27.0
pandas==2.2.2
requests==2.32.3
coinbase-advanced-py==1.8.2
PyNaCl==1.5.0
cryptography==42.0.7
pyyaml>=6.0
rich>=13.7
httpx>=0.27
python-dotenv>=1.0
pandas>=2.2
matplotlib>=3.9
plotext>=5.2
requests>=2.32
altair>=5.0

streamlit-extras>=0.4.0
